Add files via upload
A number of cryptographic functions required for the wallet and miner software, the main code makes two sets of key pairs, one for each user, and encrypts and decrypts messages that they have 'sent' to one another. Will currently only work on one computer and with one letter user names. The files that running the code will generate should be useful for understanding the formatting of messages